+++ src/components/shared/my-page/LogOut.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function LogOut() {
+  return <div className='flex text-lg font-bold text-[#999] ml-auto mr-8 mt-4'>로그아웃 </div>;
+}
