+++ src/components/shared/my-page/MyHeader.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function MyHeader() {
+  return (
+    <div className='px-6 pb-6 pt-14'>
+      <span className='text-2xl font-bold text-left text-[#1b1b1b]'>마이페이지</span>
+    </div>
+  );
+}
